dsfgdfshdfg

+++ Monopoly/src/actions/GameAction.java
@@ -47,8 +47,8 @@ public class GameAction extends HttpServlet {
 		int playerId = (int) request.getSession().getAttribute(TextKeys.playerId);
 		
 		//Spiel in HashMap finden
-		HashMap<Integer,Monopoly> gameList = (HashMap<Integer, Monopoly>) request.getServletContext().getAttribute(TextKeys.gameList);;
-		Monopoly monopoly = gameList.get(gameId);
+		HashMap<Long,Monopoly> gameList = (HashMap<Long, Monopoly>) request.getServletContext().getAttribute(TextKeys.gameList);;
+		Monopoly monopoly = gameList.get((long)gameId);
